requirements.txt
streamlit==1.32.0
pandas==2.2.0
numpy==1.26.3
matplotlib==3.8.2
wordcloud==1.9.3
docx2txt==0.8
PyPDF2==3.0.1
nltk==3.8.1
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
python-docx==1.1.0
pyparsing==3.2.1
scikit-learn==1.4.2
openai==1.12.0
python-dotenv==1.0.0
